+++ NEWS.md
@@ -1,3 +1,18 @@
+# sounDMR 2.3.2
+	- Added new feature to calculate number of cytosines included in a changepoint region
+	- Fixed a bug in creating line plots for changepoints
+
+# sounDMR 2.3.1
+        - In create_dmr_obj():
+              - Changed calculation of Percent column in LongPercent to weighted mean
+              - Fixed bug in LongMeth total_RD calculation to use sum of read depths instead of mean
+
+# sounDMR 2.3.0
+	- Removed split_by_chromosome function from workflow due to bugs 
+	- Improved notes for running whole genome analysis
+	- Improved memory issues with group_DMR()
+	- Updated figures produced by changepoint analysis
+
 # sounDMR 2.2.2
 	- Updated generate_methylframe() documentation